Allow `git push -f` to piku.
Before this change if you tried to `git push -f` it broke the update.
Git would output an error saying branches have diverged.
With this change piku basically assumes whatever you pushed into
origin/master is the thing you actually want to deploy, and uses that.
(even if it's a commit --amend or a reset to some previous point in
history etc.)