Build a single-binary deploy with pex.
This experimental PR adds the ability to build a single-binary artifact of the Piku gateway script, avoiding the need to install deps (click).
The script can run on any version of Python on any architecture because Piku does not use any compiled deps.
Test it out by running 'make' in the 'binary' subdirectory.