Add support for python deployments with uv
uv (https://docs.astral.sh/uv/) is a fast package and project manager
for python that is quickly gaining in popularity. The new deploy
functions uses uv's built in virtual environment handling, but sets the
environment path to the normal piku env path.

Since it is new, I marked it as experimental, just like the poetry one.
To not break any existing workflows, poetry is preferred if both poetry
and uv are installed.